putting TransformationProvider definition to its own file to be usable from other libs.
